Fixed bug launchpad #1382809 to manage token expired
Currenntly Freezer handle token refresh when uploading data segments,
but there's a bug in the file backup.py in the lines ~130 and ~178.
The client object called to upload the tar meta data and the
swift manifest, can  be expired as it is initialized before
the segments upload process.

launchpad: https://bugs.launchpad.net/freezer/+bug/1382809

Change-Id: I88c9d922c51a43c1722268d4dbed77e594e089aa