Restore process fails managing big data stream from swift
Currently Freezer has a bug in tar_restore() and
object_to_stream() functions, as the remote chunks needs
to be set to a far smaller value to keep the connection
stream working during restore tasks.

Also the way how the end of data stream is implemented now
is not efficient for big data streams, as it should be managed
adding False or None in the Pipe and properly managed in
the other side of the pipe.

The default segments size is now set to 64MB from 128MB.
This can be more appropriate for low resources nodes.

Change-Id: I00c4cc984e90ea1368b8842cd112a35085e14b11
LAUNCHPAD: https://bugs.launchpad.net/freezer/+bug/1384795