Containers created by Freezer will have freezer_ as prefix
The containers used by freezer to executed backups needs to have
freezer_ prefix in the name. If the user provider container doesn't
have the prefix, it is automatically added also to the container
segments name. This is done to quickly identify the containers

Added --path-to-backup to the existing ones (-F, --file-to-backup)
as can be more intuitive for the user

Bumped version to: 1.1.0 as 1.0.9-2 is now tagged in upstream

Change-Id: I891391ea07f75b5dcdc43d9b1d65a75837065f30
LAUNCHPAD: https://blueprints.launchpad.net/freezer/+spec/freezer-containers