Main loop refactor
The code relative to the various actions is moved into their
respective job objects.
A factory function creates the correct type of job object according to
the action requested.
This replaces the long "if" chain in the main loop

The main loop calls the Job object's execute() method.

Change-Id: Iec1acfe62575ba062f9b56171784d1d74dc86556
Implements: blueprint freezer-main-refactor