basic integration tests for freezer-agent
The test provided in the tests/integration directory
execute the command line version of the freezer-agent.

The tests execute backups and restores, passing the parameters
on the command line, and check the results.

Types of integration tests depend on the environment variables defined.

Local storage tests are enabled by default and use temporary
files and directories, while SSH and Swift storage testing need
also access to existing external resources.

To enable the ssh storage testing, the following environment
variables need to be defined:

 - FREEZER_TEST_SSH_USERNAME
 - FREEZER_TEST_SSH_KEY
 - FREEZER_TEST_SSH_HOST
 - FREEZER_TEST_CONTAINER

To enable the swift storage testing, the following environment
variables need to be defined:

 - FREEZER_TEST_OS_USERNAME
 - FREEZER_TEST_OS_TENANT_NAME
 - FREEZER_TEST_OS_REGION_NAME
 - FREEZER_TEST_OS_PASSWORD
 - FREEZER_TEST_OS_AUTH_URL

Tests involving LVM snapshots are executed only with superuser
privileges and if the following env variable is not defined:

 - FREEZER_TEST_NO_LVM

Change-Id: I39f0c9bf62f7c82f81ad95f19cf0f7aefb059633