Fixed deps in setup.py and a bug in tests
Currently the deps of oslo.i18n wasn't aligned with
Kilo global-requirements.

Now tests are not execute by runtests.py but tox
call directly py.test, as there was a bug in the exec
of runtests.py. This is a temporary solution as
testr will be used soon as testing framework

Close bug: 1498078

Change-Id: I07daddb14ab75fe089c98c8e00c24ecb550ada82