LVM snapshot mountpoint fix and improved usability
Solves an issue where the correct path to backup was not
found in case the lvm mount point is not the filesystem root

Snapshot logic has been improved.
Freezer is able to set the lvm parameters by itself, using
a combination of default values and values extracted from
user-provided parameters.

The user can request the use of filsystem snapshots using
the following new flag:

  --snapshot  request a filesystem snapshot

Change-Id: I46af658fbabb2fe0a78bf3398ffb86f6b711d36e