Improved freezerc and freezer-agent script management
With this patch the scripts are generated by automatically
only from setuptools. The /bin/freezerc script is removed
and the content is incorporated in freezer.main

This patch should make more portable freezer on
win32 platform.

Unfortunately in order to support win32 platforms
we need to use a version of PBR >= 1.0.0,
to support env markers. Now requirements.txt reflects that

Change-Id: I31a4a13711f8c042072e3cc8785078fdbb971cb7