Import pep3134daemon as local module
pep3134daemon is now imported from local path
rather than pip. This generated many issues
as the package is not in the global-requirements.txt
of kilo and liberty. Also pbr in the kilo release
does not support env markers which further complitated
the installation. This change needs to be backported to
stable/kilo also.

Also some minor fix for RST format.

Change-Id: I34db28ffc928703b01e6430a661214d66d81c519