Adding devstack plugin for freezer gate job
Currently the plugin is moslty a copy of the
freeer-api plugin. At first instance we can enable
the gate and have the job executign without error,
after this we can commit new changes to execute
freezer specific tests. Anyway the freezer-api plugin
is most likley needed to execute the freezer-agent and
freezer-scheduler tests

Implements bp: freezer-dsvm
Depends-on: I837e3fe973d72c792cb34711cef9f6507a004d49

Change-Id: Iee995d5b32c17c4dbee5f5736b26bf3b9aa8ca14