Fix the failure of loading the instance of MongoDbMode
It is failed to load instance of MongoDbMode when the mode is 'mongo' and
loading the mode instance at BackupJob.execute_method. Because the
mod_name=freezer.mode.mongo.MongoMode is not exist, the current
mod_name=freezer.mode.mongo.MongoDbMode. It needs to change the class name
of MongoDbMode to MongoMode

Change-Id: I687f411c98bc1841df5e60937ed87078eae405d6
Closes-Bug: #1589788