Add high-level backup utilities to BaseFreezerTest
This adds a set of new utilities to BaseFreezerTest that can create,
mutate, and automatically clean up local freezer-agent backups with a
particular set of parameters. These utilities can also be used as a
drop-in replacement for manually-created backups in existing tests.

Existing backup mutation is necessary so integration tests can create
scenarios that cannot be generated by freezer-agent directly, such as
simulating aged backups in order to test flags like
`--remove-from-date`.

Change-Id: I64760fee15c39be269223d4e7891f3ea9a8d0cb0