Always collect freezer.log when tests fail
While I330ca7b189d0d87cc8eff6c1b295a5f924bad475 fixed error reporting
for integration tests, it prevented the script from continuing to the
point that the freezer log file was copied when an error occurred.

This patch links the log to the log directory in advance to make sure
it will be uploaded to the server regardless of whether or not
devstack actually completes successfully.

Change-Id: I5c335d2a28d426693aa7379a7567272f4cee3c69