Require a backup name for backup/restore/admin
Currently, it's possible to leave out the '--backup-name' parameter
without raising an error. Due to the configured default value, this
can result in creating or searching for backups that have names ending
with a literal '_False'.

This patch sets the default value to None (to avoid converting the
False to a string) and adds a validation step to require
'--backup-name' to be set for all backup, restore, and admin actions.

Change-Id: Iebf8355b8b2f44546990c0f4b1847d4c4f82ccca
Closes-Bug: #1594952