failed to start freezer-scheduler
Edit the /etc/freezer/scheduler.conf file and
complete the following actions:
log_file = scheduler.log
log_dir = /home/tmp
use_stderr = false
no_api = true

start freezer-scheduler:
The freezer-scheduler isn't started successfully.

When start freezer-scheduler, the CONF.log_dir + CONF.log_file should
be passed to daemon.start() as parameter. Currently, the only
CONF.log_file was passed to. In the file daemon.py, the file name will
be passed to setup_logging()->configure_logging(). In the function
configure_logging(), the variable "expanded_dir_name" will be set as
"empty". So the exception was raised when creating directory to
use utils.create_dir(). The freezer-scheduler process will be aborted.

In the freezer-scheduler module, there are 2 setup_logging() funtions.
The one is in the arguments.py, the another is in the daemon.py.
The Python standard logging library is being used in the daemon.py.
We should use oslo_log instead of the logging. And the setup_logging()
in the daemon.py should be removed.

When the oslo_log is used to record the log, The freezer-scheduler
daemon can be started successfully.

Change-Id: I397cedf782d86548e7dc8bc5efc6caae9163cbd9
Partial-Bug: #1610904