Fixed the use of --insecure cli option
Fixed freezer-agent to make the correct use of --insecure cli opt.
So now freezer-agent will check OS_INSECURE if not available will
look for --insecure cli opt.
Also had to fixture freezer config opts to  Allows overriding
configuration settings for the test. So added FreezerBaseTestCase
as a Base Testcase for most of tests.


Change-Id: I4a293322c79240ff5da7b238e5f5a08594788c7e
Closes-Bug: #1614139