Use LOG.error to record log when initializing LOG
Delete the directory "scheduler" under /etc/freezer/, start
freezer-scheduler, print the error info as follow:
No handlers could be found for logger "freezer.scheduler.arguments"

In the process of initializing oslo_log, _LOG.error was used to
record error info. Before the initialization is finished, the
_LOG.error can't be used to record log info.

Use freezer_utils.create_dir() to check if the directory exist or
not, if not, will create the directory. If the error occurs, The
log info will be wrote into log file.

Change-Id: I19d7b6efd52af66142c01e017e29db3792ef0398
Closes-Bug: #1627245