Implemet PIPESTATUS env var to handle error correctly.
As tar is here piped into openssl, we only get the return code for
openssl and lose the one of tar. This patch fixes this issue using
bash PIPESTATUS variable.

The piped command now returns return code of tar if openssl succeeds.
Otherwise it returns return code of openssl. That means, both the
commands should succeed.

Change-Id: I6f9de52539f7284d9479805fc4ca478f2ffd9123
Closes-Bug: #1596587