Fix ReSizeStream
Fix issue in 'next' method of ReSizeStream class. If StopIteration
error is rasen while doing "self.reminder += self.stream.next()", it
wasn't processed in 'except StopIteration' block.

We need to fix this as it leads to error when doing backup of small images,
which size is less than ReSizeStream chunk size.

Change-Id: I21dfba3967c6a582b273f0e342ca435b14db0a0a