Force Swift storage driver to always get a new client
At the moment, the Swift storage codebase does not have a consistent
way of using the Swift client, as it switches from grabbing an
existing client and creating new ones.

Due to the architecture of how restores happen, there are several
times where the same storage instance is held by the main process,
the process which reads from the storage and the engine (to retrieve
metadata).  All of these accesses result in unexpected behaviour which
can raise SSL errors such as: DECRYPTION_FAILED_OR_BAD_RECORD_MAC

The overall cost of creating a new client for every request is very
minimal and it eliminates all possibility of multithreading issues
that occur as per suggested by the author of the requests Python
library[0], which is used by the Swift client.

This patch eliminates all calls to create_swift() indirectly and makes
sure that all accesses to the Swift client create a new instance. In
my local testing, there was no noticable performance change but
no more SSL failures.

Closes-Bug: #1682606

[0]: https://github.com/kennethreitz/requests/issues/1906#issuecomment-201238170

Change-Id: I2d43f14943fc52fd6b6cb994251c97e00d3eb863