Fix backup nova instance with creating job error
Currently, it will fail when doing nova backup with creating
a backup job. It is because freezer-scheduler will call freezer-agent
with '--config config-file', but the config opts in config-file (from
job file) were not parsed to CONF, at the same time EngineManager uses
the CONF.engine_name(now it is the default 'tar'), this will lead the
code go into wrong branch, although it will not throw errors and show
the backup is success, but when doing restore after this, we can find
that the backup data can not be used and print EOF error in log file.

This patch will fix the above issue.

Change-Id: I826e4ed71dfd2cde6103e8432faba7adb4648c36
Closes-Bug: #1710238
(cherry picked from commit 50df2c8ca3e1083ea53a0937a83a35bb9ed1d499)