Make nova tests less strict
When calling nova's server list, the list of instances aren't guaranteed
to be in a particular order, so the order in which the mocked backup
calls happens is nondeterministic and results in flaky tests. This
affects packagers who run these tests on every package build. Use
any_order=True to make the tests more flexible.

Change-Id: I85edb41e77cc66a6b8dd5c86cbefb22ed61e2700