lvm backup: calculate path_to_backup after volume is mounted
For the time being, the path_to_backup for backup with lvm snapshot
is not taken directly from --path-to-backup parameter, but adjusted
according the mount info of the corresponding volume hinted by that
parameter. However this info is inferred before the volume is actually
mounted so it tends to be inaccuate and leads to errors.

So do the calculation after volume mounting instead.

Change-Id: I0d7f3125520aa27ba1d2d94ff52e3cc0eea4916c
Closes-Bug: #1799183
Story: 2004152
Task: 27623