lvm-backup: Normalize path when comparing current path and mount point
Before umounting and removing lvm volume, we need to make sure we are
not within the mount point that is to be removed.
Previous logic lacks consideration that path can have parts like
trailing slashes, which could end up skipping changing dir when it is
needed.
This patch proposes to normalize the path before doing the check to
avoid such mistakes.

Closes-Bug: #1799334
Story: 2004153
Task: 27625
Change-Id: Iaa8c8ff523b1d29a1addeaa3d802053d63aba80a