Make the low-constraints job work and modify coverage version
Now this style can't make low-constraints job work, because the pip
will constrain a library to the first item it encounters. Then the
version defined in the uppper file will be installed here.

The solution is to put the upper file in the deps.

Example:
When running low-constraints job:
According to the low-constraints file:
Babel==2.5.3[1] should be installed
Actually the log[2] shows that Babel===2.7.0[3] was installed.

Reference Link:
[1]. https://github.com/openstack/freezer/blob/master/lower-constraints.txt#L6
[2]. http://logs.openstack.org/85/657185/2/check/openstack-tox-lower-constraints/f9d0a50/job-output.txt.gz
[3]. https://releases.openstack.org/constraints/upper/master
[4]. https://review.opendev.org/#/c/601188
[5]. http://lists.openstack.org/pipermail/openstack-discuss/2019-May/006478.html

The coverage version need >=4.5.0 to support py36:
[1]. https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=879269
[2]. https://travis-ci.org/agx/git-buildpackage/jobs/534117241

Change-Id: Idf278f4a7aca4470f5996e5166e4724259ef944b