Update Sphinx requirement
Sphinx 2.0 no longer works on python 2.7, so we need to start capping
it there as well.

The errors[1] are as follow:

Requirement(package='sphinx', location='', specifiers='!=1.6.6,!=1.6.7,>=1.6.5'
does not match "python_version>='3.4'"
Requirement(package='sphinx', location='', specifiers='!=1.6.6,!=1.6.7,>=1.6.5'
does not match "python_version=='2.7'"
Could not find a global requirements entry to match package sphinx. If the package
is already included in the global list, the name or platform markers there may not
match the local settings.

Reference:
[1]. http://logs.openstack.org/56/665856/4/check/requirements-check/e565cc2/job-output.txt.gz
[2]. https://review.opendev.org/#/c/658965/

Change-Id: Ida52a0921e4ca281954b2ac80bd0bf7f82ec17bd