Switch to the opensuse-15 nodeset
We are trying to deprecate the opensuse-150 label type in favor of the
opensuse-15 label. The reason for this is opensuse 15, like centos,
doesn't support its older point releases and instead expects users to
keep up to date on its minor updates. This allows us to carry one fewer
image in opendev and ensures we are testing supported releases.

Change-Id: Ia62f411b437d8796aec5d98c49ec59dbe736e4e3