Fix parsing config file for freezer job
Config parsers do not guess datatypes of values
in configuration files, always storing them internally
as strings. This means that we need to convert booleans, as
they are parsed as Strings from config file.

This patch is fixing typo from my previous commit where
None is needed to remove from second if clause as it is
already in first if clause and option_value should be False.

Change-Id: Id60e75b5f999b70626addaf47b4abd1ad15555d6