Fix deprecated zuul queue syntax and other gate issues
There are multiple issues with the gate. This patch combines the fixes
to unblock the gate.

1. Zuul deprecated (and removed) declaring shared queues at a pipeline
level [1]. This patch fixes the syntax to make the gate work again for
freezer.

2. Zuul CI jobs default nodeset is now Ubuntu 22.04 Jammy Jellyfish,
that does not provide anymore python2 binary, so python-dev needs to be
removed from bindep.txt

3. Latest tox does not allow to list the passed environment variables
in one line separated by space. One option is to list them in separate
lines. Also, external command allowing variable name changed.
basepython as python3 is also conflicts now and it is not necessary.
It looks like tox requires now to install now explicitly all the
requirements listed in requirements.txt.
tox.ini is updated accordingly to the above listed problems.

[1] https://lists.openstack.org/pipermail/openstack-discuss/2022-May/028603.html

Change-Id: I3c95bb48c7fdd0c752db121f1745f64a53b5a0e5