Remove unsupported distros and disable failing checks
In order to land any changes in current state it makes sense to
disable CI first, as it was not passing for a while now.

With that we can land some changes that will fix Zuul configuration
errors for the project.

Change-Id: Iec3a054fa9aaca4af670e6c5e5ae760d13558f46
(cherry picked from commit 735302b27dc520b754a87644a35c23a767ed7784)